+++ prequ/utils.py
@@ -26,7 +26,7 @@ pip_version_info = tuple(safeint(digit) for digit in pip.__version__.split('.'))
 def assert_compatible_pip_version():
     # Make sure we're using a reasonably modern version of pip
     if not pip_version_info >= (8, 0):
-        print('Pip Requ requires at least version 8.0 of pip ({} found), '
+        print('Prequ requires at least version 8.0 of pip ({} found), '
               'perhaps run `pip install --upgrade pip`?'.format(pip.__version__))
         sys.exit(4)